Python/task2_tBot/requirements.txt
 aiogram==3.*
 python-dotenv==1.0.1
 requests==2.32.3